ingestion/ingestion-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
httpx>=0.27.0
pypdf>=4.2.0
python-docx>=1.1.0
chromadb>=0.5.0
openai>=1.30.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
tiktoken>=0.7.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
tenacity>=8.3.0
pyyaml>=6.0.1
spacy>=3.8.0
redis[asyncio]>=5.0.0
